Fix Linux compile

+++ Sources/HodEngine/Editor/InspectorWindow.cpp
@@ -32,6 +32,8 @@
 #include "HodEngine/Game/PrefabResource.hpp"
 #include "HodEngine/Game/Prefab.hpp"
 
+#include <cmath>
+
 namespace hod::editor
 {
 	DECLARE_WINDOW_DESCRIPTION(InspectorWindow, "Inspector", true)
@@ -259,7 +261,7 @@ namespace hod::editor
 					ImVec2 buttonPos = ImGui::GetCursorScreenPos();
 					ImVec2 mousePos = ImGui::GetMousePos();
 
-					float distance = std::sqrt(std::powf(buttonPos.x - mousePos.x, 2) + std::powf(buttonPos.y - mousePos.y, 2));
+					float distance = std::sqrt(std::pow(buttonPos.x - mousePos.x, 2) + std::pow(buttonPos.y - mousePos.y, 2));
 					distance = std::clamp(distance, 15.0f, 40.0f);
 					float alpha = 1.0f - ((distance - 15.0f) / (40.0f - 15.0f));
